cli/requirements.txt
click==8.0.1
pyserial==3.5
